+++ README.md
@@ -1,7 +1,15 @@
-# Good TURN
+# VK/WB TURN Proxy
 [English version](README.en.md)
 
-Проброс трафика WireGuard/Hysteria через TURN сервера VK звонков или Яндекс телемоста. Пакеты шифруются DTLS 1.2, затем параллельными потоками через TCP или UDP отправляются на TURN сервер по протоколу STUN ChannelData. Оттуда по UDP отправляются на ваш сервер, где расшифровываются и передаются в WireGuard. Логин/пароль от TURN генерируются из ссылки на звонок.
+Проброс трафика WireGuard/Hysteria через TURN сервера VK звонков или WB Stream. Пакеты шифруются DTLS 1.2, затем параллельными потоками через TCP или UDP отправляются на TURN сервер по протоколу STUN ChannelData. Оттуда по UDP отправляются на ваш сервер, где расшифровываются и передаются в WireGuard. Логин/пароль от TURN генерируются из ссылки на звонок.
+
+## Возможности
+
+- **VK Calls** — TURN credentials от VK API с автоматическим решением капчи (Not Robot)
+- **WB Stream** — TURN credentials от WB Stream API (LiveKit ICE)
+- **Кеширование** — 10 минут с общим кешем на 4 потока
+- **DTLS-обфускация** — обход DPI через DTLS-туннель
+- **Множественные соединения** — до N параллельных подключений к TURN
 
 **Обновление: Многопользовательский прокси-сервер**
 Текущая реализация поддерживает одновременную работу нескольких пользователей через один прокси-сервер.
@@ -11,18 +19,34 @@
 - **Балансировка:** Исходящий трафик от сервера к клиенту распределяется между всеми активными DTLS-потоками пользователя (Round-Robin).
 
 Только для учебных целей!
+
+## Структура клиента
+
+```
+client/
+├── credentials.go    # Общий кеш credentials, сериализация запросов
+├── vk.go             # VK API: Token 1→4 chain, HTTP-запросы
+├── vk_captcha.go     # VK Captcha: PoW solving, Not Robot flow
+├── wb.go             # WB Stream: guest register → room → LiveKit ICE
+└── main.go           # DTLS/TURN соединения, CLI, основной цикл
+```
+
 ## Настройка
+
 Нам понадобится:
 1. Ссылка на действующий ВК звонок: создаём свой (нужен аккаунт вк), или гуглим `"https://vk.com/call/join/"`.
-Ссылка действительна вечно, если не нажимать "завершить звонок для всех"
-2. Или ссыска на звонок Яндекс телемоста: `"https://telemost.yandex.ru/j/"`. Её лучше не гуглить, так как видно подключение к конференции
-3. VPS с установленным WireGuard
-4. Для андроида: скачать Termux из F-Droid
+   Ссылка действительна вечно, если не нажимать "завершить звонок для всех"
+2. VPS с установленным WireGuard
+3. Для андроида: скачать Termux из F-Droid
+
 ### Сервер
+
 ```
 ./server -listen 0.0.0.0:56000 -connect 127.0.0.1:<порт wg>
 ```
+
 ### Клиент
+
 #### Android
 
 **Рекомендуемый способ:**
@@ -30,7 +54,8 @@
 
 **Альтернативный способ (через Termux):**
 - В клиентском конфиге WireGuard меняем адрес сервера на `127.0.0.1:9000`, ставим MTU 1280
--  **Добавляем Termux в исключения WireGuard. Нажимаем "сохранить".**
+- **Добавляем Termux в исключения WireGuard. Нажимаем "сохранить".**
+
 В Termux:
 ```
 termux-wake-lock
@@ -44,20 +69,27 @@ termux-wake-unlock
 cp /sdcard/Download/client-android ./
 chmod 777 ./client-android
 ```
-Запускаем:
+
+**VK режим:**
 ```
 ./client-android -peer <ip сервера wg>:56000 -vk-link <VK ссылка> -listen 127.0.0.1:9000
 ```
+
+**WB режим:**
+```
+./client-android -wb -peer <ip сервера wg>:56000 -listen 127.0.0.1:9000
+```
+
 Дополнительные флаги:
 - `-session-id <hex>`: установить фиксированный ID сессии (32 символа hex).
+- `-n <число>`: количество подключений к TURN (по умолчанию 4).
+- `-udp`: использовать UDP для TURN (по умолчанию TCP).
+- `-turn <ip>`: переопределить адрес TURN сервера.
+- `-port <port>`: переопределить порт TURN сервера.
+- `-no-dtls`: без DTLS-обфускации (может привести к бану).
 
-Или
-```
-./client-android -udp -turn 5.255.211.241 -peer <ip сервера wg>:56000 -yandex-link <Ya ссылка> -listen 127.0.0.1:9000
-```
-
-**Если после включения VPN в терминале вылезают ошибки DNS, попробуйте в Wireguard включить VPN только для нужных приложений.**
 #### Linux
+
 В клиентском конфиге WireGuard меняем адрес сервера на `127.0.0.1:9000`, ставим MTU 1280
 
 Скрипт будет добавлять маршруты к нужным ip:
@@ -67,11 +99,13 @@ chmod 777 ./client-android
 ```
 
 ```
-./client-linux -udp -turn 5.255.211.241 -peer <ip сервера wg>:56000 -yandex-link <Ya ссылка> -listen 127.0.0.1:9000 | sudo routes.sh
+./client-linux -wb -peer <ip сервера wg>:56000 -listen 127.0.0.1:9000 | sudo routes.sh
 ```
 
 Не включайте впн, пока программа не установит соединение! В отличие от андроида, здесь часть запросов будет идти через впн (dns и запрос подключения к turn)
+
 #### Windows
+
 В клиентском конфиге WireGuard меняем адрес сервера на `127.0.0.1:9000`, ставим MTU 1280
 
 В PowerShell от Администратора (чтобы скрипт прописывал маршруты):
@@ -81,40 +115,42 @@ chmod 777 ./client-android
 ```
 
 ```
-./client.exe -udp -turn 5.255.211.241 -peer <ip сервера wg>:56000 -yandex-link <Ya ссылка> -listen 127.0.0.1:9000 | routes.ps1
+./client.exe -wb -peer <ip сервера wg>:56000 -listen 127.0.0.1:9000 | routes.ps1
 ```
 
-Не включайте впн, пока программа не установит соединение! В отличие от андроида, здесь часть запросов будет идти через впн (dns и запрос подключения к turn)
+Не включайте впн, пока программа не установит соединение! В отличие от андродида, здесь часть запросов будет идти через впн (dns и запрос подключения к turn)
+
 ### Если не работает
-С помощью опции `-turn` можно указать адрес TURN сервера вручную. Это должен быть сервер ВК, Макса или Одноклассников (ссылка вк) или Яндекса (ссылка яндекса). Возможно потом составлю список.
+
+С помощью опции `-turn` можно указать адрес TURN сервера вручную. Это должен быть сервер ВК, Макса или Одноклассников (ссылка вк) или WB Stream (режим -wb).
 
 Если не работает TCP, попробуйте добавить флаг `-udp`.
 
 Добавьте флаг `-n 1` для более стабильного подключения в 1 поток (ограничение 5 Мбит/с для ВК)
 
-## Яндекс телемост
-**UPD. ТЕЛЕМОСТ ЗАКРЫЛИ**
+## VK Auth Flow
 
-В отличие от ВК, сервера яндекса не ограничивают скорость, так что по умолчанию стоит `-n 1`. Увеличение этого числа может привести к временной блокировке по IP из-за переполнения конференции фейковыми участниками.
+1. **Token 1** — анонимный токен (`login.vk.ru`)
+2. **getCallPreview** — превью звонка (опционально)
+3. **Token 2** — анонимный токен для звонка (`api.vk.ru`)
+   - При капче → PoW solving → retry
+4. **Token 3** — OK session key (`calls.okcdn.ru`)
+5. **Token 4** — TURN credentials (`calls.okcdn.ru`)
 
-В режиме `-udp` скорость обычно больше
+## WB Auth Flow
 
-Большинство диапазонов IP TURN серверов Яндекса не работают, указывайте вручную через `-turn`
-<details>
-    <summary>
-        Рабочие IP
-    </summary>
+1. **Guest register** — регистрация гостя (`stream.wb.ru`)
+2. **Create room** — создание комнаты
+3. **Join room** — подключение к комнате
+4. **Get token** — получение roomToken
+5. **LiveKit ICE** — WebSocket к LiveKit, protobuf парсинг TURN
 
+## Кеширование
 
-    5.255.211.241
-    5.255.211.242
-    5.255.211.243
-    5.255.211.245
-    5.255.211.246
-
-
-</details>
-Спасибо https://github.com/KillTheCensorship/Turnel за часть кода :)
+- TTL: **10 минут** (safety margin 60 секунд)
+- Один кеш на **4 потока** (`streamID / 4`)
+- Fast path через `RLock`
+- Сериализация fetch через глобальный `fetchMu`
 
 ## v2ray
 
@@ -225,4 +261,9 @@ chmod 777 ./client-android
 </details>
 
 ## Direct mode
-С флагом `-no-dtls` можно отправлять пакеты без обфускации DTLS и подключаться к обычным серверам Wireguard. Может привести к бану от вк/яндекса.
+
+С флагом `-no-dtls` можно отправлять пакеты без обфускации DTLS и подключаться к обычным серверам Wireguard. Может привести к бану от ВК/WB.
+
+Спасибо https://github.com/KillTheCensorship/Turnel за часть кода :)
+
+Функционал WB Stream основан на проекте https://github.com/jaykaiperson/lionheart
